+++ src/platform/windows/mod.rs
@@ -1,6 +1,7 @@
 #![cfg(target_os = "windows")]
 
 use winapi;
+use winapi::shared::windef::HWND;
 
 pub use self::events_loop::{EventsLoop, EventsLoopProxy};
 pub use self::monitor::MonitorId;
@@ -8,14 +9,14 @@ pub use self::window::Window;
 
 #[derive(Clone, Default)]
 pub struct PlatformSpecificWindowBuilderAttributes {
-    pub parent: Option<winapi::HWND>,
+    pub parent: Option<HWND>,
 }
 
 unsafe impl Send for PlatformSpecificWindowBuilderAttributes {}
 unsafe impl Sync for PlatformSpecificWindowBuilderAttributes {}
 
 // TODO: document what this means
-pub type Cursor = *const winapi::wchar_t;
+pub type Cursor = *const winapi::ctypes::wchar_t;
 
 // Constant device ID, to be removed when this backend is updated to report real device IDs.
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -23,7 +24,7 @@ pub struct DeviceId;
 const DEVICE_ID: ::DeviceId = ::DeviceId(DeviceId);
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct WindowId(winapi::HWND);
+pub struct WindowId(HWND);
 unsafe impl Send for WindowId {}
 unsafe impl Sync for WindowId {}